Clean up Search UI
