+++ src/android/GetGrinnected/MyApplication/app/src/main/java/screens/SearchInProgressCode.kt
@@ -1,0 +1,170 @@
+package screens
+
+// This is code that was made for trying to implement a search bar using search bar tutorials
+// It doesn't work but would be a better implementation of search bar
+/*
+@SuppressLint("UnrememberedMutableState")
+@OptIn(ExperimentalMaterial3Api::class)
+@Composable
+fun SearchScreen(viewModel: SearchViewModel) {
+    val searchQuery = viewModel.searchQuery
+    val eventEntities by AppRepository.events
+    // Converts them to event data type
+    val events = eventEntities.map { it.toEvent() }
+    val searchResults by viewModel.searchResults.collectAsStateWithLifecycle()
+    SearchScreen(
+        searchQuery = searchQuery,
+        searchResults = events,
+        onSearchQueryChange = {viewModel.onSearchQueryChange(it) }
+    )
+}
+
+@RequiresApi(Build.VERSION_CODES.TIRAMISU)
+@OptIn(ExperimentalMaterial3Api::class)
+@Composable
+fun SearchScreen(
+    searchQuery: String,
+    searchResults: List<Event>,
+    onSearchQueryChange: (String) -> Unit
+) {
+    Box(
+        modifier = Modifier
+            .fillMaxSize()
+            .semantics { isTraversalGroup = true }
+    ) {
+        SearchBar(
+            modifier = Modifier
+                .align(Alignment.TopCenter)
+                .semantics { traversalIndex = 0f },
+            query = searchQuery,
+            onQueryChange = onSearchQueryChange,
+            onSearch = {},
+            placeholder = {
+                Text(text = "Search Events")
+            },
+            leadingIcon = {
+                Icon(
+                    imageVector = Icons.Default.Search,
+                    tint = MaterialTheme.colorScheme.onSurface,
+                    contentDescription = null
+                )
+            },
+            trailingIcon = {
+                if (searchQuery.isNotEmpty()) {
+                    IconButton(onClick = { onSearchQueryChange("") }) {
+                        Icon(
+                            imageVector = Icons.Default.Close,
+                            tint = MaterialTheme.colorScheme.onSurface,
+                            contentDescription = "Clear search"
+                        )
+                    }
+                }
+            },
+            content = {
+                LazyColumn(
+                verticalArrangement = Arrangement.spacedBy(32.dp),
+                contentPadding = PaddingValues(16.dp),
+                modifier = Modifier.fillMaxSize()
+            ) {
+                items(
+                    count = searchResults.size,
+                    key = { index -> searchResults[index].eventid },
+                    itemContent = { index ->
+                        val event = searchResults[index]
+                        EventCard(event = event)
+                    }
+                )
+            }},
+            active = true,
+            onActiveChange = {},
+            tonalElevation = 0.dp
+        )
+}}
+
+@Composable
+fun EventListItem(
+    event: Event,
+    modifier: Modifier = Modifier
+) {
+    Row(
+        horizontalArrangement = Arrangement.SpaceBetween,
+        modifier = modifier.fillMaxWidth()
+    ) {
+        Text(text = event.event_name)
+        event.event_date?.let { Text(text = it) }
+        event.event_time?.let { Text(text = it) }
+    }
+}
+*/
+
+/*
+@OptIn(ExperimentalMaterial3Api::class)
+@Composable
+fun CustomizableSearchBar(
+    query: String,
+    onQueryChange: (String) -> Unit,
+    onSearch: (String) -> Unit,
+    searchResults: MutableList<Event>,
+    onResultClick: (String) -> Unit,
+    // Customization options
+    placeholder: @Composable () -> Unit = { Text("Search") },
+    leadingIcon: @Composable (() -> Unit)? = { Icon(Icons.Default.Search, contentDescription = "Search") },
+    trailingIcon: @Composable (() -> Unit)? = null,
+    supportingContent: (@Composable (String) -> Unit)? = null,
+    leadingContent: (@Composable () -> Unit)? = null,
+    @SuppressLint("ModifierParameter") modifier: Modifier = Modifier
+) {
+    // Track expanded state of search bar
+    var expanded = false
+
+    Box(
+        modifier
+            .fillMaxSize()
+            .semantics { isTraversalGroup = true }
+    ) {
+        SearchBar(
+            modifier = Modifier
+                .align(Alignment.TopCenter)
+                .semantics { traversalIndex = 0f },
+            inputField = {
+                // Customizable input field implementation
+                SearchBarDefaults.InputField(
+                    query = query,
+                    onQueryChange = onQueryChange,
+                    onSearch = {
+                        onSearch(query)
+                        expanded = false
+                    },
+                    expanded = expanded,
+                    onExpandedChange = { expanded = it },
+                    placeholder = placeholder,
+                    leadingIcon = leadingIcon,
+                    trailingIcon = trailingIcon
+                )
+            },
+            expanded = expanded,
+            onExpandedChange = { expanded = it },
+        ) {
+            // Show search results in a lazy column for better performance
+            LazyColumn {
+                items(count = searchResults.size) { index ->
+                    val resultText = searchResults[index]
+                    ListItem(
+                        headlineContent = { Text(resultText.toString()) },
+                        supportingContent = supportingContent?.let { { it(resultText.toString()) } },
+                        leadingContent = leadingContent,
+                        colors = ListItemDefaults.colors(containerColor = Color.Transparent),
+                        modifier = Modifier
+                            .clickable {
+                                onResultClick(resultText.toString())
+                                expanded = false
+                            }
+                            .fillMaxWidth()
+                            .padding(horizontal = 16.dp, vertical = 4.dp)
+                    )
+                }
+            }
+        }
+    }
+}
+*/
